Диплом 3я часть.Дополнительная функциональность - комментарии/ после проверки
