+++ ewm-service/src/main/java/ru/practicum/ewmservice/comment/dto/CommentDto.java
@@ -1,5 +1,6 @@
 package ru.practicum.ewmservice.comment.dto;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import ru.practicum.ewmservice.event.model.Event;
@@ -16,9 +17,11 @@ import java.time.LocalDateTime;
 public class CommentDto {
     Long id;
     String text;
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime createdtime;
     Event event;
     User author;
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime editedtime;
     Boolean isdeleted;
 }
