+++ ewm-service/src/main/java/ru/practicum/ewmservice/comment/validation/Create.java
@@ -1,0 +1,4 @@
+package ru.practicum.ewmservice.comment.validation;
+
+public interface Create {
+}
