+++ ewm-service/src/main/java/ru/practicum/ewmservice/comment/validation/Update.java
@@ -1,0 +1,4 @@
+package ru.practicum.ewmservice.comment.validation;
+
+public interface Update {
+}
